Module bpaf::_documentation

Project documentation

See official website for more up to date version.

Introduction and design goals - A quick intro. What, why and how

Tutorials - practical, learning oriented guides

HOWTO - practical, oriented to solving problems guides

Parsing cookbook - How to parse less frequent combinations

Theory explanation - Theoretical information about abstractions used by the library, oriented for understanding

Modules

_0_intro

_1_tutorials

_2_howto

_3_cookbook

_4_explanation