Module bpaf::_documentation::_1_tutorials::_0_types_of_arguments::_0_switch

↑ Types of arguments ↑

Option arguments or arguments →

Options, switches or flags

Options or flags usually starts with a dash, a single dash for short options and a double dash for long one. Several short options can usually be squashed together with a single dash in front of them to save on typing: -vvv can be parsed the same as -v -v -v. Options don’t have any other information apart from being there or not. Relative position usually does not matter and --alpha --beta should parse the same as --beta --alpha.

$ cargo --help
$ ls -la
$ ls --time --reverse

Combinatoric example

#[derive(Debug, Clone)]
pub struct Options {
    verbose: bool,
    release: bool,
    default_features: bool,
}

pub fn options() -> OptionParser<Options> {
    let verbose = short('v')
        .long("verbose")
        .help("Produce verbose output")
        .switch();
    let release = long("release")
        .help("Build artifacts in release mode")
        .flag(true, false);
    let default_features = long("no-default-features")
        .help("Do not activate default features")
        // default_features uses opposite values,
        // producing `true` when value is absent
        .flag(false, true);

    construct!(Options {
        verbose,
        release,
        default_features,
    })
    .to_options()
}

fn main() {
    println!("{:?}", options().run())
}

Derive example

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Options {
    /// Produce verbose output
    // bpaf uses `switch` for `bool` fields in named
    // structs unless consumer attribute is present.
    // But it is also possible to give it explicit
    // consumer annotation to serve as a reminder:
    // #[bpaf(short, long, switch)]
    #[bpaf(short, long)]
    verbose: bool,

    #[bpaf(flag(true, false))]
    /// Build artifacts in release mode
    release: bool,

    /// Do not activate default features
    // default_features uses opposite values,
    // producing `true` when value is absent
    #[bpaf(long("no-default-features"), flag(false, true))]
    default_features: bool,
}

fn main() {
    println!("{:?}", options().run())
}

Output

In --help output bpaf shows switches as usual flags with no meta variable attached

$ app --help

Usage: app [-v] [--release] [--no-default-features]

Available options:

-v, --verbose

Produce verbose output

--release

Build artifacts in release mode

--no-default-features

Do not activate default features

-h, --help

Prints help information

Both switch and flag succeed if value is not present, switch returns true, flag returns second value.

$ app
Options { verbose: false, release: false, default_features: true }

When value is present - switch returns true, flag returns first value.

$ app --verbose --no-default-features --detailed
Error: --detailed is not expected in this context

Like with most parsrs unless specified switch and flag consume at most one item from the command line:

$ app --no-default-features --no-default-features
Error: argument --no-default-features cannot be used multiple times in this context

For more detailed info see NamedArg::switch and NamedArg::flag

↑ Types of arguments ↑

Option arguments or arguments →