bpaf::_documentation::_1_tutorials::_0_types_of_arguments
Module _4_exotic

← Commands or subcommands

↑ Types of arguments ↑

Exotic schemas

While modern software tends to use just the options listed above you can still encounter programs created before those options became the norm and they use something completely different, let me give a few examples, see the parsing cookbook about actually parsing them

su takes an option that consists of a single dash -

$ su -

find considers everything between --exec and ; to be a single item. this example calls ls -l on every file find finds.

$ find /etc --exec ls -l '{}' \;

Xorg and related tools use flag-like items that start with a single + to enable a feature and with - to disable it.

$ xorg -backing +xinerama

dd takes several key-value pairs, this would create a 100M file

$ dd if=/dev/zero of=dummy.bin bs=1M count=100

Most of the command line arguments in Turbo C++ 3.0 start with /. For example, option /x tells it to use all available extended memory, while /x[=n] limits it to n kilobytes

C:\PROJECT>TC /x=200

← Commands or subcommands

↑ Types of arguments ↑