Module bpaf::_documentation::_1_tutorials::_1_combinatoric_api::_0_simple_parser::_1_argument

← Switch parser

↑ Making a simple parser ↑

Positional item parser →

Argument parser

Next in complexity would be a parser to consume a named argument, such as -p my_crate. Same as with the switch parser it starts from a NamedArg but the next method is NamedArg::argument. This method takes a metavariable name - a short description that will be used in the --help output. rustc also needs to know the parameter type you are trying to parse, there are several ways to do it:

fn simple_argument_1() -> impl Parser<u32> {
    // rustc figures out the type from returned value
    long("number").argument("NUM")
}
 
fn simple_argument_2() -> impl Parser<String> {
    // type is specified explicitly with turbofish
    long("name").argument::<String>("NAME")
}
 
fn file_parser() -> OptionParser<PathBuf> {
    // OptionParser is a type for finalized parser, at this point you can
    // start adding extra information to the `--help` message
    long("file").argument::<PathBuf>("FILE").to_options()
}

You can use any type for as long as it implements FromStr. To parse items that don’t implement it you can first parse a String or OsString and then use Parser::parse, see the next chapter on how to do that.

Full example with some sample inputs and outputs:

use bpaf::*;

pub fn options() -> OptionParser<usize> {
    short('s')
        .long("size")
        .help("Defines size of an object")
        .argument::<usize>("SIZE")
        .to_options()
}

fn main() {
    println!("{:?}", options().run())
}

Output

By default all arguments are required so running with no arguments produces an error

$ app
Error: expected --size=SIZE, pass --help for usage information

Bpaf accepts various combinations of names and adjacencies:

$ app -s100
100

$ app --size 300
300

$ app -s=42
42

$ app --size=14
14

Since not every string is a valid number - bpaf would report any parsing failures to the user directly

$ app --size fifty
Error: couldn't parse fifty: invalid digit found in string

In addition to the switch you defined bpaf generates switch for user help which will include the description from the help method

$ app --help

Usage: app -s=SIZE

Available options:

-s, --size=SIZE

Defines size of an object

-h, --help

Prints help information

← Switch parser

↑ Making a simple parser ↑

Positional item parser →