Module bpaf::_documentation::_1_tutorials::_2_derive_api::_1_custom_names

← Getting started with derive macro

↑ Derive API tutorial ↑

Customizing the consumers →

Customizing flag and argument names

By default names for flag names are taken directly from the field names so usually you don’t have to do anything about it, but you can change it with annotations on the fields themselves:

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Options {
    /// A custom switch
    #[bpaf(short, long)]
    switch: bool,

    /// A custom argument
    #[bpaf(long("my-argument"), short('A'))]
    argument: usize,
}

fn main() {
    println!("{:?}", options().run())
}

Output

bpaf uses custom names in help message

$ app --help

Usage: app [-s] -A=ARG

Available options:

-s, --switch

A custom switch

-A, --my-argument=ARG

A custom argument

-h, --help

Prints help information

As well as accepts them on a command line and uses in error message

$ app --switch
Error: expected --my-argument=ARG, pass --help for usage information

$ app -A 42 -s
Options { switch: true, argument: 42 }

Rules for picking the name are:

With no annotations field name longer than a single character becomes a long name, single character name becomes a short name

Adding either long or short disables item 1, so adding short disables the long name

long or short annotation without a parameter derives a value from a field name

long or short with a parameter uses that instead

You can have multiple long and short annotations, the first of each type becomes a visible name, remaining are used as hidden aliases

And if you decide to add names - they should go to the left side of the annotation list

← Getting started with derive macro

↑ Derive API tutorial ↑

Customizing the consumers →