Module bpaf::_documentation::_1_tutorials::_2_derive_api::_2_custom_consumers

← Customizing flag and argument names

↑ Derive API tutorial ↑

Transforming parsed values →

Customizing the consumers

By default, bpaf picks parsers depending on a field type according to those rules:

bool fields are converted into switches: NamedArg::switch

() (unit) fields, unit variants of an enum or unit structs themselves are handled as NamedArg::req_flag and thus users must always specify them for the parser to succeed

All other types with no Vec/Option are parsed using FromStr, but smartly, so non-utf8 PathBuf/OsString are working as expected.

For values wrapped in Option or Vec bpaf derives the inner parser and then applies applies logic from Parser::optional and Parser::many respectively.

You can change it with annotations like switch, argument or positional

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Options {
    /// A custom switch
    #[bpaf(short, switch)]
    switch: bool,

    /// Custom number
    #[bpaf(positional("NUM"))]
    argument: usize,
}

fn main() {
    println!("{:?}", options().run())
}

Output

bpaf generates help message with a short name only as described

$ app --help

Usage: app [-s] NUM

Available positional items:

NUM

Custom number

Available options:

-s

A custom switch

-h, --help

Prints help information

And accepts the short name only

$ app -s 42
Options { switch: true, argument: 42 }

long name is missing

$ app --switch 42
Error: --switch is not expected in this context

With arguments that consume a value you can specify its type using turbofish-line syntax

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Options {
    /// A custom argument
    #[bpaf(positional::<usize>("LENGTH"))]
    argument: usize,
}
 
fn main() {
    let opts = options().run();
    println!("{:?}", opts)
}

← Customizing flag and argument names

↑ Derive API tutorial ↑

Transforming parsed values →