Module bpaf::_documentation::_3_cookbook::_01_dd

← find(1): find -exec commands -flags terminated by \;

↑ Parsing cookbook ↑

Xorg(1): Xorg +xinerama +extension name →

dd(1): dd if=/dev/zero of=/dev/null bs=1000

This example implements syntax similar to dd command. The main idea is to implement something to make it simple to make parsers for PREFIX=SUFFIX, where prefix is fixed for each parser - for example if= or of= and suffix is parsed with usual FromStr trait.

The function tag serves this purpose. It performs the following steps:

consume any item that starts with a prefix at any argument position with any and ParseAny::anywhere

attaches help message and custom metadata to make --help friendlier

parses suffix with Parser::parse

The rest of the parser simply uses tag to parse a few of dd arguments

examples/dd.rs

//! This is not a typical bpaf usage,
//! but you should be able to replicate command line used by dd
use bpaf::{any, construct, doc::Style, short, OptionParser, Parser};
use std::str::FromStr;

#[derive(Debug, Clone)]
#[allow(dead_code)]
pub struct Options {
    magic: bool,
    in_file: String,
    out_file: String,
    block_size: usize,
}

/// Parses a string that starts with `name`, returns the suffix parsed in a usual way
fn tag<T>(name: &'static str, meta: &str, help: &'static str) -> impl Parser<T>
where
    T: FromStr,
    <T as std::str::FromStr>::Err: std::fmt::Display,
{
    // it is possible to parse OsString here and strip the prefix with
    // `os_str_bytes` or a similar crate
    any("", move |s: String| Some(s.strip_prefix(name)?.to_owned()))
        // this defines custom metavar for the help message
        .metavar(&[(name, Style::Literal), (meta, Style::Metavar)][..])
        .help(help)
        .anywhere()
        .parse(|s| s.parse())
}

fn in_file() -> impl Parser<String> {
    tag::<String>("if=", "FILE", "read from FILE")
        .fallback(String::from("-"))
        .display_fallback()
}

fn out_file() -> impl Parser<String> {
    tag::<String>("of=", "FILE", "write to FILE")
        .fallback(String::from("-"))
        .display_fallback()
}

fn block_size() -> impl Parser<usize> {
    // it is possible to parse notation used by dd itself as well,
    // using usuze only for simplicity
    tag::<usize>("bs=", "SIZE", "read/write SIZE blocks at once")
        .fallback(512)
        .display_fallback()
}

pub fn options() -> OptionParser<Options> {
    let magic = short('m')
        .long("magic")
        .help("a usual switch still works")
        .switch();
    construct!(Options {
        magic,
        in_file(),
        out_file(),
        block_size(),
    })
    .to_options()
}

fn main() {
    println!("{:#?}", options().run());
}

Output

bpaf generates usual help message with

$ app --help

Usage: app [-m] [if=FILE] [of=FILE] [bs=SIZE]

Available options:

-m, --magic

a usual switch still works

if=FILE

read from FILE

[default: -]

of=FILE

write to FILE

[default: -]

bs=SIZE

read/write SIZE blocks at once

[default: 512]

-h, --help

Prints help information

Unlike usual application dd takes it arguments in shape of operations KEY=VAL without any dashes, plus usual --help and --version flags.

To handle that we define custom basic parsers that make handling such operations easy

$ app if=/dev/zero of=/tmp/blob bs=1024
Options { magic: false, in_file: "/dev/zero", out_file: "/tmp/blob", block_size: 1024 }

← find(1): find -exec commands -flags terminated by \;

↑ Parsing cookbook ↑

Xorg(1): Xorg +xinerama +extension name →