Function bpaf::batteries::get_usage

pub fn get_usage<T>(parser: OptionParser<T>) -> String
where
    T: Debug,

Get usage for a parser

In some cases you might want to print usage if user gave no command line options, in this case you should add an enum variant to a top level enum, make it hidden with #[bpaf(hide)], make it default for the top level parser with something like #[bpaf(fallback(Arg::Help))].

When handling cases you can do something like this for Help variant:

    ...
    Arg::Help => {
        println!("{}", get_usage(parser()));
        std::process::exit(0);
    }
    ...