bpaf::doc

Struct Doc

pub struct Doc { /* private fields */ }

String with styled segments.

You can add style information to generated documentation and help messages For simpliest possible results you can also pass a string slice in all the places that require impl Into<Doc>

Implementations

impl Doc

pub fn monochrome(&self, full: bool) -> String

Render a monochrome version of the document

full indicates if full message should be rendered, this makes difference for rendered help message, otherwise you can pass true.

impl Doc

pub fn render_markdown(&self, full: bool) -> String

Render doc into markdown document, used by documentation sample generator

impl Doc

pub fn text(&mut self, text: &str)

Append a fragment of plain text to Doc

See Doc for usage examples

pub fn literal(&mut self, text: &str)

Append a fragment of literal text to Doc

See Doc for usage examples

pub fn emphasis(&mut self, text: &str)

Append a fragment of text with emphasis to Doc

See Doc for usage examples

pub fn invalid(&mut self, text: &str)

Append a fragment of unexpected user input to Doc

See Doc for usage examples

pub fn meta(&mut self, meta: MetaInfo<'_>, for_usage: bool)

Append a fragment of parser metadata to Doc

See Doc for usage examples

pub fn doc(&mut self, buf: &Doc)

Append a Doc to Doc

See Doc for usage examples

pub fn em_doc(&mut self, buf: &Doc)

Append a Doc to Doc for plaintext documents try to format first line as a help section header