bpaf

Enum ParseFailure

pub enum ParseFailure {
    Stdout(Doc, bool),
    Completion(String),
    Stderr(Doc),
}

Unsuccessful command line parsing outcome, use it for unit tests

When OptionParser::run_inner produces Err(ParseFailure) it means that the parser couldnโt produce the value it supposed to produce and the program should terminate.

If you are handling variants manually - Stdout contains formatted output and you can use any logging framework to produce the output, Completion should be printed to stdout unchanged - shell completion mechanism relies on that. In both cases application should exit with error code of 0. Stderr variant indicates a genuinly parsing error which should be printed to stderr or a logging framework of your choice as an error and the app should exit with error code of 1. ParseFailure::exit_code is a helper method that performs printing and produces the exit code to use.

For purposes of for unit testing for user parsers, you can consume it with ParseFailure::unwrap_stdout and ParseFailure::unwrap_stdout - both of which produce a an unformatted String that parser might produce if failure type is correct or panics otherwise.

Variants

Stdout(Doc, bool)

Print this to stdout and exit with success code

Completion(String)

This also goes to stdout with exit code of 0, this cannot be Doc because completion needs more control about rendering

Stderr(Doc)

Print this to stderr and exit with failure code

Implementations

impl ParseFailure

pub fn unwrap_stderr(self) -> String

Returns the contained stderr values - for unit tests

Panics

Panics if failure contains stdout

pub fn unwrap_stdout(self) -> String

Returns the contained stdout values - for unit tests

Panics

Panics if failure contains stderr

pub fn exit_code(self) -> i32

Returns the exit code for the failure

pub fn print_message(&self, max_width: usize)

Prints a message to stdout or stderr appropriate to the failure.