bpaf

Module error

Structs

Error

Unsuccessful command line parsing outcome, internal representation

MissingItem

Missing item in a context

Enums

Message 🔒

ParseFailure

Unsuccessful command line parsing outcome, use it for unit tests

Functions

check_conflicts 🔒

only_once 🔒

summarize_missing 🔒

go over all the missing items, pick the left most scope

textual_part 🔒