Function bpaf::command

pub fn command<T>(name: &'static str, subparser: OptionParser<T>) -> Command<T>where
    T: 'static,

Subcommand parser

Subcommands allow to use a totally independent parser inside a current one. Inner parser can have its own help message, description, version and so on. You can nest them arbitrarily too.

Alternatively you can create commands using command

Important restriction

When parsing command arguments from command lines you should have parsers for all your named values and command before parsers for positional items. In derive API fields parsed as positional should be at the end of your struct/enum. Same rule applies to parsers with positional fields or commands inside: such parsers should go to the end as well.

Use check_invariants in your test to ensure correctness.

For example for non positional non_pos and a command command parsers

let valid = construct!(non_pos(), command());
let invalid = construct!(command(), non_pos());

bpaf panics during help generation unless if this restriction holds

Combinatoric use

Structure Command you get by calling this method is a builder that allows to add additional aliases with short, long (only first short and first long names are visible to --help) and override help. Without help override bpaf would use first line from the description

#[derive(Debug, Clone)]
enum Cmd {
    Check {
        workspace: bool,
    }
};

// First of all you need an inner parser
fn check_workspace() -> OptionParser<Cmd> {
    // Define a parser to use in a subcommand in a usual way.
    // This parser accepts a single --workspace switch
    let workspace = long("workspace")
        .help("Check all packages in the workspace")
        .switch();

    // and attach some meta information to it in a usual way
    construct!(Cmd::Check { workspace })
        .to_options()
        // description to use for command's help
        .descr("Check a package for errors")
}

// Convert subparser into a parser.
fn check_workspace_command() -> impl Parser<Cmd> {
    command("check", check_workspace())
        // help to use to list the command
        .help("Check a package command")
}

Derive usage

#[derive(Clone, Debug, Bpaf)]
enum Cmd {
    #[bpaf(command)]
    /// Check a package command
    Check {
        /// Check all the packages in the workspace
        workspace: bool
    }
}

Example

$ app --help
// displays global help, not listed in this example
$ app check --help
// displays help for check: "Check a package command"
$ app check
// Cmd::Check(CheckWorkspace(false))
$ app check --workspace
// Cmd::Check(CheckWorkspace(true))