bpaf::params

Struct ParseCommand

pub struct ParseCommand<T> {
    pub(crate) longs: Vec<&'static str>,
    pub(crate) shorts: Vec<char>,
    pub(crate) help: Option<Doc>,
    pub(crate) subparser: OptionParser<T>,
    pub(crate) adjacent: bool,
}

Builder structure for the [command]

Created with [command], implements parser for the inner structure, gives access to help.

Fields
longs: Vec<&'static str>shorts: Vec<char>help: Option<Doc>subparser: OptionParser<T>adjacent: bool
Implementations

impl<P> ParseCommand<P>

pub fn help<M>(self, help: M) -> Self
where M: Into<Doc>,

Add a brief description to a command

bpaf uses this description along with the command name in help output so it shouldn’t exceed one or two lines. If help isn’t specified bpaf falls back to descr from the inner parser.

Combinatoric usage

fn inner() -> OptionParser<bool> {
    short('i')
        .help("Mysterious inner switch")
        .switch()
        .to_options()
        .descr("performs an operation")
}

fn mysterious_parser() -> impl Parser<bool> {
    inner().command("mystery")
        .help("This command performs a mystery operation")
}

Derive usage

bpaf_derive uses doc comments for inner parser, no specific options are available. See descr for more details

/// This command performs a mystery operation
#[derive(Debug, Clone, Bpaf)]
#[bpaf(command)]
struct Mystery {
    #[bpaf(short)]
    /// Mysterious inner switch
    inner: bool,
}

Example

$ app --help
    <skip>
Available commands:
    mystery  This command performs a mystery operation

pub fn short(self, short: char) -> Self

Add a custom short alias for a command

Behavior is similar to short, only first short name is visible.

pub fn long(self, long: &'static str) -> Self

Add a custom hidden long alias for a command

Behavior is similar to long, but since you had to specify the first long name when making the command - this one becomes a hidden alias.

pub fn adjacent(self) -> Self

Allow for the command to succeed even if there are non consumed items present

Normally a subcommand parser should handle the rest of the unconsumed elements thus allowing only “vertical” chaining of commands. adjacent modifier lets command parser to succeed if there are leftovers for as long as all comsumed items form a single adjacent block. This opens possibilities to chain commands sequentially.

Let’s consider two examples with consumed items marked in bold :

cmd -a -b -c -d

cmd -a -c -b -d

In the first example -b breaks the adjacency for all the consumed items so parsing will fail, while here in the second one the name and all the consumed items are adjacent to each other so parsing will succeed.

Combinatoric example

#[derive(Debug, Clone)]
pub struct Options {
    premium: bool,
    commands: Vec<Cmd>,
}

#[derive(Debug, Clone)]
// shape of the variants doesn't really matter, let's use all of them :)
enum Cmd {
    Eat(String),
    Drink { coffee: bool },
    Sleep { time: usize },
}

fn cmd() -> impl Parser<Cmd> {
    let eat = positional::<String>("FOOD")
        .to_options()
        .descr("Performs eating action")
        .command("eat")
        .adjacent()
        .map(Cmd::Eat);

    let coffee = long("coffee")
        .help("Are you going to drink coffee?")
        .switch();
    let drink = construct!(Cmd::Drink { coffee })
        .to_options()
        .descr("Performs drinking action")
        .command("drink")
        .adjacent();

    let time = long("time").argument::<usize>("HOURS");
    let sleep = construct!(Cmd::Sleep { time })
        .to_options()
        .descr("Performs taking a nap action")
        .command("sleep")
        .adjacent();

    construct!([eat, drink, sleep])
}

pub fn options() -> OptionParser<Options> {
    let premium = short('p')
        .long("premium")
        .help("Opt in for premium serivces")
        .switch();
    let commands = cmd().many();
    construct!(Options { premium, commands }).to_options()
}

fn main() {
    println!("{:?}", options().run())
}

Derive example

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Options {
    #[bpaf(short, long)]
    /// Opt in for premium serivces
    pub premium: bool,
    #[bpaf(external(cmd), many)]
    pub commands: Vec<Cmd>,
}

#[derive(Debug, Clone, Bpaf)]
pub enum Cmd {
    #[bpaf(command, adjacent)]
    /// Performs eating action
    Eat(#[bpaf(positional("FOOD"))] String),
    #[bpaf(command, adjacent)]
    /// Performs drinking action
    Drink {
        /// Are you going to drink coffee?
        coffee: bool,
    },
    #[bpaf(command, adjacent)]
    /// Performs taking a nap action
    Sleep {
        #[bpaf(argument("HOURS"))]
        time: usize,
    },
}

fn main() {
    println!("{:?}", options().run())
}

Output

Example implements a parser that supports one of three possible commands:

$ app --help

Usage: app [-p] [COMMAND ...]...

Available options:

-p, --premium

Opt in for premium serivces

-h, --help

Prints help information

Available commands:

eat

Performs eating action

drink

Performs drinking action

sleep

Performs taking a nap action

As usual every command comes with its own help

$ app drink --help

Performs drinking action

Usage: app drink [--coffee]

Available options:

--coffee

Are you going to drink coffee?

-h, --help

Prints help information

Normally you can use one command at a time, but making commands adjacent lets parser to succeed after consuming an adjacent block only and leaving leftovers for the rest of the parser, consuming them as a Vec<Cmd> with many allows to chain multiple items sequentially

$ app eat Fastfood drink --coffee sleep --time=5
Options { premium: false, commands: [Eat("Fastfood"), Drink { coffee: true }, Sleep { time: 5 }] }

The way this works is by running parsers for each command. In the first iteration eat succeeds, it consumes eat fastfood portion and appends its value to the resulting vector. Then second iteration runs on leftovers, in this case it will be drink --coffee sleep --time=5. Here drink succeeds and consumes drink --coffee portion, then sleep parser runs, etc.

You can mix chained commands with regular arguments that belong to the top level parser

$ app sleep --time 10 --premium eat 'Bak Kut Teh' drink
Options { premium: true, commands: [Sleep { time: 10 }, Eat("Bak Kut Teh"), Drink { coffee: false }] }

But not inside the command itself since values consumed by the command are not going to be adjacent

$ app sleep --time 10 eat --premium 'Bak Kut Teh' drink
Error: expected FOOD, pass --help for usage information

impl<T> ParseCommand<T>

fn item(&self) -> Item