bpaf::parsers

Struct ParseCollect

pub struct ParseCollect<P, C, T> {
    pub(crate) inner: P,
    pub(crate) catch: bool,
    pub(crate) ctx: PhantomData<(C, T)>,
}

Apply inner parser several times and collect results into FromIterator, created with collect, Implements catch

Fields
inner: Pcatch: boolctx: PhantomData<(C, T)>
Implementations

impl<T, C, P> ParseCollect<P, C, T>

pub fn catch(self) -> Self

Handle parse failures

Can be useful to decide to skip parsing of some items on a command line When parser succeeds - catch version would return a value as usual if it fails - catch would restore all the consumed values and return None.

There’s several structures that implement this attribute: ParseOptional, ParseMany and ParseSome, behavior should be identical for all of them.

Combinatoric example

#[derive(Debug, Clone)]
pub struct Options {
    height: Vec<usize>,
    height_str: Vec<String>,
    width: Vec<usize>,
    width_str: Vec<String>,
}

pub fn options() -> OptionParser<Options> {
    // contains catch
    let height = long("height")
        .help("Height of a rectangle")
        .argument::<usize>("PX")
        .some("You must specify some heights")
        .catch();

    let height_str = long("height").argument::<String>("PX").many().hide();

    // contains no catch
    let width = long("width")
        .help("Width of a rectangle")
        .argument::<usize>("PX")
        .some("You must specify some widths");

    let width_str = long("width").argument::<String>("PX").many().hide();

    construct!(Options {
        height,
        height_str,
        width,
        width_str
    })
    .to_options()
}

fn main() {
    println!("{:?}", options().run())
}

Derive example

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Options {
    #[bpaf(long, argument("PX"), some("You must specify some heights"), catch)]
    /// Height of a rectangle
    height: Vec<usize>,

    #[bpaf(long("height"), argument("PX"), many, hide)]
    height_str: Vec<String>,

    #[bpaf(long, argument("PX"), some("You must specify some widths"))]
    /// Width of a rectangle
    width: Vec<usize>,

    #[bpaf(long("width"), argument("PX"), many, hide)]
    width_str: Vec<String>,
}

fn main() {
    println!("{:?}", options().run())
}

Output

Despite parser producing a funky value - help looks like you would expect from a parser that takes two values

$ app --help

Usage: app --height=PX... --width=PX...

Available options:

--height=PX

Height of a rectangle

--width=PX

Width of a rectangle

-h, --help

Prints help information

When executed with no parameters parser fails because some requires you to specify at least one matching parameter

$ app
Error: You must specify some heights

When executed with expected parameters fields with usize get their values

$ app --height 100 --width 100 --height 12 --width 44
Options { height: [100, 12], height_str: [], width: [100, 44], width_str: [] }

With incorrect value for --height parameter inner part of height parser fails, some combined with catch handles this failure and produces [] without consuming value from the command line. Parser height_str runs next and consumes the value as a string

$ app --height 10 --height twenty --width 33
Options { height: [10], height_str: ["twenty"], width: [33], width_str: [] }

In case of wrong --width - parser width fails, parser for some sees this as a “value is present but not correct” and propagates the error outside, execution never reaches width_str parser

$ app --height 10 --width 33 --width ten
Error: couldn't parse ten: invalid digit found in string