Struct bpaf::parsers::ParseMany

pub struct ParseMany<P> { /* private fields */ }

Apply inner parser several times and collect results into Vec, created with many, implements catch.

Implementations

impl<P> ParseMany<P>

pub fn catch(self) -> Self

Handle parse failures

Can be useful to decide to skip parsing of some items on a command line When parser succeeds - catch version would return a value as usual if it fails - catch would restore all the consumed values and return None.

There’s several structures that implement this attribute: ParseOptional, ParseMany and ParseSome, behavior should be identical for all of them.

Combinatoric usage

#[derive(Debug, Clone)]
pub struct Options {
    version: Option<usize>,
    feature: Option<String>,
}
pub fn options() -> OptionParser<Options> {
    let version = positional::<usize>("VERS").optional().catch();
    let feature = positional::<String>("FEAT").optional();
    construct!(Options { version, feature }).to_options()
}

Derive usage

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Options {
    #[bpaf(positional("VERS"), catch)]
    version: Option<usize>,
    #[bpaf(positional("FEAT"), catch)]
    feature: Option<String>,
}

Examples

No argument, both parsers succeed due to optional alone

% app 
Options { version: None, feature: None }

Decimal value - version parser succeeds, second parser returns None due to optional

% app 10
Options { version: Some(10), feature: None }

String value - version parser fails, catch handles that, second parser handles still available "feature"

% app feature
Options { version: None, feature: Some("feature") }

Two values are present, both parsers succeed handles still available "feature"

% app 1000 feature
Options { version: Some(1000), feature: Some("feature") }