Struct bpaf::Named

pub struct Named { /* private fields */ }

A named thing used to create flag, switch or argument

Create it with short or long.

Ways to consume data

bpaf supports several different ways user can specify values on a command line:

flag - a string that consists of two dashes (--flag) and a name and a single dash and a single character (-f) - long and short name respectively. Depending on it present or absent from the command line primitive flag parser takes one of two values. User can combine several short flags in a single invocation: -a -b -c is the same as -abc.

$ app -a -bc

switch - similar to flag, but instead of custom values bpaf uses bool. switch mostly serves as a convenient alias to .flag(true, false)

$ app -a -bc

argument - a short or long flag followed by either a space or = and then by a string literal. -f foo, --flag bar or -o=- are all valid argument examples. Note, string literal can’t start with - unless separated from the flag with = and should be valid utf8 only. To consume OsString encoded values you can use argument_os.

$ app -o file.txt

positional - an arbitrary utf8 string literal (that can’t start with -) passed on a command line, there’s also positional_os variant that deals with OsString named. Usually represents input files such as cat file.txt, but can serve other purposes.

$ cat file.txt

command - a fixed utf8 string literal that starts a separate subparser that only gets executed when command name is present. For example cargo build invokes command "build" and after "build" cargo starts accepting values it won’t accept otherwise

$ cargo build --out-dir my_target
// works since command "build" supports --out-dir argument
$ cargo check --out-dir my_target
// fails since --out-dir isn't a valid argument for command "check"

As most of the other parsers bpaf treats anything to the right of -- symbol as positional arguments regardless their names:

$ app -o file.txt -- --those --are --positional --items

Combinatoric usage

Named items (argument, flag and switch) can have up to 2 visible names (short and long) and multiple hidden short and long aliases if needed. It’s also possible to consume items from environment variables using env. You usually start with short or long function, then apply short / long / env / help repeatedly to build a desired set of names then transform it into a parser using flag, switch or positional.

fn an_item() -> impl Parser<String> {
    short('i')
        .long("item")
        .long("also-item") // but hidden
        .env("ITEM")
        .help("A string used by this example")
        .argument("ITEM")
}

Example

$ app --help
    <skip>
    -i --item <ITEM>  [env:ITEM: N/A]
                      A string used by this example
    <skip>

Derive usage

Unlike combinatoric API where you forced to specify names for your parsers derive API allows to omit some or all the details:

If no naming information is present at all - bpaf_derive would use field name as a long name (or a short name if field name consists of a single character)

If short or long annotation is present without an argument - bpaf_derive would use first character or a full name as long and short name respectively. It won’t try to add implicit long or short name from the previous item.

If short or long annotation is present with an argument - those are values bpaf_derive would use instead of the original field name

If env(arg) annotation is present - bpaf_derive would generate .env(arg) method:

const DB: &str = "top_secret_database";

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Config {
   /// flag with no annotation
   pub flag_1: bool,

   /// explicit short suppresses long
   #[bpaf(short)]
   pub flag_2: bool,

   /// explicit short with custom letter
   #[bpaf(short('z'))]
   pub flag_3: bool,

   /// explicit short and long
   #[bpaf(short, long)]
   pub deposit: bool,

   /// implicit long + env variable from DB constant
   #[bpaf(env(DB))]
   pub database: String,

   /// implicit long + env variable "USER"
   #[bpaf(env("USER"))]
   pub user: String,
}

Example

$ app --help
   <skip>
        --flag-1         flag with no annotation
   -f                    explicit short suppresses long
   -z                    explicit short with custom letter
   -d, --deposit         explicit short and long
       --database <ARG>  [env:top_secret_database: N/A]
                         implicit long + env variable from DB constant
       --user <ARG>      [env:USER = "pacak"]
                         implicit long + env variable "USER"
   <skip>

Implementations

impl Named

pub fn short(self, short: char) -> Self

Add a short name to a flag/switch/argument

You can specify it multiple times, bpaf would use items past the first one as hidden aliases.

fn parse_bool() -> impl Parser<bool> {
    short('f')
        .short('F')
        .long("flag")
        .help("a flag that does a thing")
        .switch()
}

See Named for more details

pub fn long(self, long: &'static str) -> Self

Add a long name to a flag/switch/argument

You can specify it multiple times, bpaf would use items past the first one as hidden aliases.

fn parse_bool() -> impl Parser<bool> {
    short('f')
        .long("flag")
        .long("Flag")
        .help("a flag that does a thing")
        .switch()
}

See Named for more details

pub fn env(self, variable: &'static str) -> Self

Environment variable fallback

If named value isn’t present - try to fallback to this environment variable. You can specify it multiple times, bpaf would use items past the first one as hidden aliases.

fn parse_string() -> impl Parser<String> {
    short('k')
        .long("key")
        .env("API_KEY")
        .help("Use this API key to access the API")
        .argument("KEY")
}

See Named and env for more details and examples

pub fn help<M>(self, help: M) -> Selfwhere
M: Into<String>,

Add a help message to a flag/switch/argument

Combinatoric usage

fn parse_bool() -> impl Parser<bool> {
    short('f')
        .long("flag")
        .help("a flag that does a thing")
        .switch()
}

Derive usage

bpaf_derive converts doc comments into option help by following those rules:

It skips blank lines, if present.

It stops parsing after a double blank line.

#[derive(Debug, Clone, Bpaf)]
struct Options {
    /// This line is part of help message
    ///
    /// So is this one
    ///
    ///
    /// But this one isn't
    key: String,
}

See Named for more details

pub fn switch(self) -> impl Parser<bool>

Simple boolean flag

Parser produces true if flag is present in a command line or false otherwise

fn parse_bool() -> impl Parser<bool> {
    short('f')
        .long("flag")
        .help("a flag that does a thing")
        .switch()
}

See Named for more details

pub fn flag<T>(self, present: T, absent: T) -> impl Parser<T>where
T: Clone + 'static,

Flag with custom present/absent values

Combinatoric usage

Parser produces present if flag is present in a command line or absent otherwise

#[derive(Clone)]
enum Flag {
    Absent,
    Present,
}
fn parse_flag() -> impl Parser<Flag> {
    short('f')
        .long("flag")
        .help("a flag that does a thing")
        .flag(Flag::Present, Flag::Absent)
}

Derive usage

Currently available only with external annotation

#[derive(Debug, Clone)]
enum Flag {
    Absent,
    Present,
}

fn flag() -> impl Parser<Flag> {
    short('f')
        .long("flag")
        .help("a flag that does a thing")
        .flag(Flag::Present, Flag::Absent)
}

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(external)]
    pub flag: Flag,
}

See Named for more details

pub fn req_flag<T>(self, present: T) -> impl Parser<T>where
T: Clone + 'static,

Required flag with custom value

Parser produces a value if present and fails otherwise. Not very useful by itself and works best in combination with other parsers.

Combinatoric usage

#[derive(Clone)]
enum Decision {
    On,
    Off,
    Undecided
}

// user can specify either --on or --off, parser would fallback to `Undecided`
fn parse_decision() -> impl Parser<Decision> {
    let on = long("on").req_flag(Decision::On);
    let off = long("off").req_flag(Decision::Off);
    on.or_else(off).fallback(Decision::Undecided)
}

Example

$ app --on
// Decision::On
$ app
// Decision::Undecided

// counts number of flags `-v` on the command line
fn verbosity() -> impl Parser<usize> {
    short('v').req_flag(()).many().map(|v| v.len())
}

Example

$ app
// 0
$ app -vv -v
// 3

Derive usage

bpaf would transform field-less enum variants into values combined by req_flag. In addition to naming annotations (short, long and env) such variants also accept hide and default annotations. Former hides it from --help (see hide, later makes it a default choice if preceeding variants fail to parse. You shoud only use default annotation on the last variant of enum.

#[derive(Debug, Clone, Bpaf)]
enum Decision {
    On,
    Off,
    #[bpaf(long, hide, default)]
    Undecided,
}

See Named for more details

pub fn argument(self, metavar: &'static str) -> impl Parser<String>

Named argument in utf8 (String) encoding

Argument must contain only valid utf8 characters. For OS specific encoding see argument_os.

Combinatoric usage

fn parse_string() -> impl Parser<String> {
    short('n')
        .long("name")
        .argument("NAME")
}

Derive usage

bpaf_derive would automatically pick between argument and argument_os depending on a field type but you can specify it manually to override the metavar value

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short('n'), argument("NAME"))]
    name: String,
}

See Named for more details

pub fn argument_os(self, metavar: &'static str) -> impl Parser<OsString>

Named argument in OS specific encoding

If you prefer to panic on non utf8 encoding see argument.

Combinatoric usage

fn parse_osstring() -> impl Parser<std::ffi::OsString> {
    short('n')
        .long("name")
        .argument_os("NAME")
}

Derive usage

bpaf_derive would automatically pick between argument and argument_os depending on a field type but you can specify it manually to override the metavar value

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short('n'), argument_os("NAME"))]
    name: std::ffi::OsString,
}

See Named for more details