Struct bpaf::Positional

pub struct Positional<T> { /* private fields */ }

Parse a positional item, created with positional and positional_os

You can add extra information to positional parsers with help and strict on this struct.

Implementations

impl<T> Positional<T>

pub fn help<M>(self, help: M) -> Selfwhere
M: Into<String>,

Add a help message to a positional/positional_os parser

Combinatoric usage

fn parse_name() -> impl Parser<String> {
    positional("NAME")
        .help("a flag that does a thing")
}

Derive usage

bpaf_derive converts doc comments into option help by following those rules:

It skips blank lines, if present.

It stops parsing after a double blank line.

#[derive(Debug, Clone, Bpaf)]
struct Options (
    /// This line is part of help message
    ///
    /// So is this one
    ///
    ///
    /// But this one isn't
    String,
);

See also Named::help

pub fn strict(self) -> Self

Changes positional parser to be “strict” positional

Usually positional items can appear anywhere on a command line:

$ ls -d bpaf
$ ls bpaf -d

here ls takes a positional item bpaf and a flag -d

But in some cases it might be useful to have a stricter separation between positonal items and flags, such as passing arguments to a subprocess:

$ cargo run --example simple -- --help

here cargo takes a --help as a positional item and passes it to the example

bpaf allows to require user to pass -- for positional items with strict annotation. bpaf would display such positional elements differently in usage line as well. If your application requires several different strict positional elements - it’s better to place this annotation only to the first one.

Example

Usage line for a cargo-run like app that takes an application and possibly many strictly positional child arguments can look like this:

$ app --help
Usage: [-p SPEC] [[--bin NAME] | [--example NAME]] [--release] [<BIN>] -- <CHILD_ARG>...
<skip>

Combinatoric usage

fn options() -> impl Parser<Vec<std::ffi::OsString>> {
    positional_os("OPTS")
        .strict()
        .many()
}

Derive usage

Not available at the moment