Crate bpaf

Lightweight and flexible command line argument parser with derive and combinatoric style API

Derive and combinatoric API

bpaf supports both combinatoric and derive APIs and it’s possible to mix and match both APIs at once. Both APIs provide access to mostly the same features, some things are more convenient to do with derive (usually less typing), some - with combinatoric (usually maximum flexibility and reducing boilerplate structs). In most cases using just one would suffice. Whenever possible APIs share the same keywords and overall structure. Documentation is shared and contains examples for both combinatoric and derive style.

bpaf supports dynamic shell completion for bash, zsh, fish and elvish.

Quick links

Derive tutorial

Combinatoric tutorial

Some very unusual cases

Applicative functors? What is it all about

Batteries included

Q&A

Quick start - combinatoric and derive APIs

Derive style API, click to expand

Add bpaf under [dependencies] in your Cargo.toml

[dependencies]
bpaf = { version = "0.7", features = ["derive"] }

Define a structure containing command line attributes and run generated function

use bpaf::Bpaf;

#[derive(Clone, Debug, Bpaf)]
#[bpaf(options, version)]
/// Accept speed and distance, print them
struct SpeedAndDistance {
    /// Speed in KPH
    speed: f64,
    /// Distance in miles
    distance: f64,
}

fn main() {
    // #[derive(Bpaf)] generates `speed_and_distance` function
    let opts = speed_and_distance().run();
    println!("Options: {:?}", opts);
}

Try to run the app

% very_basic --help
Accept speed and distance, print them

Usage: --speed ARG --distance ARG

Available options:
        --speed <ARG>     Speed in KPH
        --distance <ARG>  Distance in miles
    -h, --help            Prints help information
    -V, --version         Prints version information

% very_basic --speed 100
Expected --distance ARG, pass --help for usage information

% very_basic --speed 100 --distance 500
Options: SpeedAndDistance { speed: 100.0, distance: 500.0 }

% very_basic --version
Version: 0.5.0 (taken from Cargo.toml by default)

You can check the derive tutorial for more detailed information.

Combinatoric style API, click to expand

Add bpaf under [dependencies] in your Cargo.toml

[dependencies]
bpaf = "0.7"

Declare parsers for components, combine them and run it

use bpaf::{construct, long, Parser};
#[derive(Clone, Debug)]
struct SpeedAndDistance {
    /// Dpeed in KPH
    speed: f64,
    /// Distance in miles
    distance: f64,
}

fn main() {
    // primitive parsers
    let speed = long("speed")
        .help("Speed in KPG")
        .argument::<f64>("SPEED");

    let distance = long("distance")
        .help("Distance in miles")
        .argument::<f64>("DIST");

    // parser containing information about both speed and distance
    let parser = construct!(SpeedAndDistance { speed, distance });

    // option parser with metainformation attached
    let speed_and_distance
        = parser
        .to_options()
        .descr("Accept speed and distance, print them");

    let opts = speed_and_distance.run();
    println!("Options: {:?}", opts);
}

Try to run the app

% very_basic --help
Accept speed and distance, print them

Usage: --speed ARG --distance ARG

Available options:
        --speed <ARG>     Speed in KPH
        --distance <ARG>  Distance in miles
    -h, --help            Prints help information
    -V, --version         Prints version information

% very_basic --speed 100
Expected --distance ARG, pass --help for usage information

% very_basic --speed 100 --distance 500
Options: SpeedAndDistance { speed: 100.0, distance: 500.0 }

% very_basic --version
Version: 0.5.0 (taken from Cargo.toml by default)

You can check the combinatoric tutorial for more detailed information.

Design goals: flexibility, reusability, correctness

Library allows to consume command line arguments by building up parsers for individual arguments and combining those primitive parsers using mostly regular Rust code plus one macro. For example it’s possible to take a parser that requires a single floating point number and transform it to a parser that takes several of them or takes it optionally so different subcommands or binaries can share a lot of the code:

// a regular function that doesn't depend on any context, you can export it
// and share across subcommands and binaries
fn speed() -> impl Parser<f64> {
    long("speed")
        .help("Speed in KPH")
        .argument::<f64>("SPEED")
}

// this parser accepts multiple `--speed` flags from a command line when used,
// collecting results into a vector
fn multiple_args() -> impl Parser<Vec<f64>> {
    speed().many()
}

// this parser checks if `--speed` is present and uses value of 42.0 if it's not
fn with_fallback() -> impl Parser<f64> {
    speed().fallback(42.0)
}

At any point you can apply additional validation or fallback values in terms of current parsed state of each subparser and you can have several stages as well:

#[derive(Clone, Debug)]
struct Speed(f64);
fn speed() -> impl Parser<Speed> {
    long("speed")
        .help("Speed in KPH")
        .argument::<f64>("SPEED")

        // You can perform additional validation with `parse` and `guard` functions
        // in as many steps as required.
        // Before and after next two applications the type is still `impl Parser<f64>`
        .guard(|&speed| speed >= 0.0, "You need to buy a DLC to move backwards")
        .guard(|&speed| speed <= 100.0, "You need to buy a DLC to break the speed limits")

        // You can transform contained values, next line gives `impl Parser<Speed>` as a result
        .map(|speed| Speed(speed))
}

Library follows parse, don’t validate approach to validation when possible. Usually you parse your values just once and get the results as a Rust struct/enum with strict types rather than a stringly typed hashmap with stringly typed values in both combinatoric and derive APIs.

Design goals: restrictions

The main restricting library sets is that you can’t use parsed values (but not the fact that parser succeeded or failed) to decide how to parse subsequent values. In other words parsers don’t have the monadic strength, only the applicative one - for more detailed explanation see Applicative functors? What is it all about.

To give an example, you can implement this description:

Program takes one of --stdout or --file flag to specify the output target, when it’s --file program also requires -f attribute with the filename

But not this one:

Program takes an -o attribute with possible values of 'stdout' and 'file', when it’s 'file' program also requires -f attribute with the filename

This set of restrictions allows bpaf to extract information about the structure of the computations to generate help, dynamic completion and overall results in less confusing enduser experience

bpaf performs no parameter names validation, in fact having multiple parameters with the same name is fine and you can combine them as alternatives and performs no fallback other than fallback. You need to pay attention to the order of the alternatives inside the macro: parser that consumes the left most available argument on a command line wins, if this is the same - left most parser wins. So to parse a parameter --test that can be both switch and argument you should put the argument one first.

You must place positional items at the end of a structure in derive API or consume them as last arguments in derive API.

Dynamic shell completion

bpaf implements shell completion to allow to automatically fill in not only flag and command names, but also argument and positional item values.

Enable autocomplete feature:

bpaf = { version = "0.7", features = ["autocomplete"] }

Decorate argument and positional parsers with complete to autocomplete argument values

Depending on your shell generate appropriate completion file and place it to whereever your shell is going to look for it, name of the file should correspond in some way to name of your program. Consult manual for your shell for the location and named conventions:

bash: for the first bpaf completion you need to install the whole script

$ your_program --bpaf-complete-style-bash >> ~/.bash_completion

but since the script doesn’t depend on a program name - it’s enough to do this for each next program

echo "complete -F _bpaf_dynamic_completion your_program" >> ~/.bash_completion

zsh: note _ at the beginning of the filename

$ your_program --bpaf-complete-style-zsh > ~/.zsh/_your_program

fish

$ your_program --bpaf-complete-style-fish > ~/.config/fish/completions/your_program.fish

elvish - not sure where to put it, documentation is a bit cryptic

$ your_program --bpaf-complete-style-elvish

Restart your shell - you need to done it only once or optionally after bpaf major version upgrade: generated completion files contain only instructions how to ask your program for possible completions and don’t change even if options are different.

Generated scripts rely on your program being accessible in $PATH

Design non goals: performance

Library aims to optimize for flexibility, reusability and compilation time over runtime performance which means it might perform some additional clones, allocations and other less optimal things. In practice unless you are parsing tens of thousands of different parameters and your app exits within microseconds - this won’t affect you. That said - any actual performance related problems with real world applications is a bug.

More examples

You can find a more examples here: https://github.com/pacak/bpaf/tree/master/examples

They’re usually documented or at least contain an explanation to important bits and you can see how they work by cloning the repo and running

$ cargo run --example example_name

Testing your own parsers

You can test your own parsers to maintain compatibility or simply checking expected output with run_inner

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Options {
    pub user: String
}

#[test]
fn test_my_options() {
    let help = options()
        .run_inner(Args::from(&["--help"]))
        .unwrap_err()
        .unwrap_stdout();
    let expected_help = "\
Usage --user <ARG>
<skip>
";

    assert_eq!(help, expected_help);
}

Cargo features

derive: adds a dependency on bpaf_derive crate and reexport Bpaf derive macro. You need to enable it to use derive API. Disabled by default.

extradocs: used internally to include tutorials to https://docs.rs/bpaf, no reason to enable it for local development unless you want to build your own copy of the documentation (https://github.com/rust-lang/cargo/issues/8905). Disabled by default.

batteries: helpers implemented with public bpaf API. Enabled by default.

autocomplete: enables support for shell autocompletion. Disabled by default.

bright-color, dull-color: use more colors when printing --help and such. Enabling either color feature adds some extra dependencies and might raise MRSV. If you are planning to use this feature in a published app - it’s best to expose them as feature flags:

[features]
bright-color = ["bpaf/bright-color"]
dull-color = ["bpaf/dull-color"]

Disabled by default.

Modules

_applicative

Applicative functors? What is it about?

_combinatoric_tutorial

Using the library in combinatoric style

_derive_tutorial

Using the library in derive style

_unusual

Some of the more unusual examples

batteries

Batteries included - helpful parsers that use only public API

params

Tools to define primitive parsers

parsers

This module exposes parsers that accept further configuration with builder pattern

Macros

construct

Compose several parsers to produce a single result

Structs

Args

All currently present command line parameters, use it for unit tests and manual parsing

OptionParser

Ready to run Parser with additional information attached

Enums

CompleteDecor

Various complete options decorations

ParseFailure

Unsuccessful command line parsing outcome, use it for unit tests

Traits

Parser

Simple or composed argument parser

Functions

any

Take next unconsumed item on the command line as raw String or OsString

command

Subcommand parser

env

Environment variable fallback

fail

Fail with a fixed error message

long

A flag/switch/argument that has a long name

positional

Positional argument in utf8 (String) encoding

pure

Wrap a value into a Parser

pure_with

Wrap a calculated value into a Parser

short

A flag/switch/argument that has a short name

Derive Macros

Bpaf

Derive macro for bpaf command line parser