bpaf::_documentation::_1_tutorials::_1_combinatoric_api::_0_simple_parser

Module _0_switch

↑ Making a simple parser ↑

Argument parser →

Switch parser

Let’s start with the simplest possible one - a simple switch that gets parsed into a bool.

First of all - the switch needs a name - you can start with short or long and add more names if you want: long("simple") or short('s').long("simple"). This gives something with the type NamedArg:

use bpaf::parsers::NamedArg;
fn simple_switch() -> NamedArg {
    short('s').long("simple")
}

From NamedArg you make a switch parser by calling NamedArg::switch. Usually, you do it right away without assigning NamedArg to a variable.

fn simple_switch() -> impl Parser<bool> {
    short('s').long("simple").switch()
}

The switch parser we just made implements trait Parser and to run it you convert it to OptionParser with Parser::to_options and run it with OptionParser::run

Full example with some sample inputs and outputs:

use bpaf::*;

pub fn options() -> OptionParser<bool> {
    let simple = short('s').long("simple").switch();
    simple.to_options()
}

fn main() {
    println!("{:?}", options().run())
}

Output

If you run the app with no parameters - switch will parse as false

$ app
false

Both short and long names produce true

$ app -s
true

$ app --simple
true

In addition to the switch you defined bpaf generates switch for user help

$ app --help

Usage: app [-s]

Available options:

-s, --simple

-h, --help

Prints help information

With NamedArg::help you can attach a help message that will be used in --help output.

↑ Making a simple parser ↑

Argument parser →