bpaf::_documentation::_1_tutorials::_1_combinatoric_api

Module _2_combining

← Transforming parsers

↑ Combinatoric API ↑

Subcommand parsers →

Combining multiple simple parsers

A single-item option parser can only get you so far. Fortunately, you can combine multiple parsers with construct! macro.

For sequential composition (all the fields must be present) you write your code as if you are constructing a structure, enum variant or a tuple and wrap it with construct!. Both a constructor and parsers must be present in the scope. If instead of a parser you have a function that creates one - just add () after the name:

struct Options {
    alpha: usize,
    beta: usize
}
 
fn alpha() -> impl Parser<usize> {
    long("alpha").argument("ALPHA")
}
 
fn both() -> impl Parser<Options> {
    let beta = long("beta").argument("BETA");
    // call `alpha` function, and use result to make parser
    // for field `alpha`, use parser `beta` for field `beta`
    construct!(Options { alpha(), beta })
}

Full example:

use bpaf::*;

#[derive(Debug, Clone)]
pub struct Options {
    argument: Vec<u32>,
    switches: Vec<bool>,
}

pub fn options() -> OptionParser<Options> {
    let argument = long("argument")
        .help("important argument")
        .argument("ARG")
        .many();
    let switches = long("switch").help("some switch").switch().many();
    construct!(Options { argument, switches }).to_options()
}

fn main() {
    println!("{:?}", options().run())
}

Output

Help message describes all the parser combined

$ app --help

Usage: app [--argument=ARG]... [--switch]...

Available options:

--argument=ARG

important argument

--switch

some switch

-h, --help

Prints help information

And users can pass any combinations of options, resulting parser will handle them as long as they are valid

$ app --argument 10 --argument 20
Options { argument: [10, 20], switches: [false] }

$ app --switch
Options { argument: [], switches: [true] }

$ app --switch --switch --argument 20
Options { argument: [20], switches: [true, true] }

If you are using positional parsers - they must go to the right-most side and will run in the order you specify them. For named parsers order affects only the --help message.

The second type of composition construct! offers is a parallel composition. You pass multiple parsers that produce the same result type in [] and bpaf selects one that fits best with the data user gave.

use bpaf::*;

pub fn options() -> OptionParser<f64> {
    let miles = long("miles").help("Distance in miles").argument("MI");
    let km = long("kilo").help("Distance in kilometers").argument("KM");
    construct!([miles, km]).to_options()
}

fn main() {
    println!("{:?}", options().run())
}

Output

Help message describes all the parser combined

$ app --help

Usage: app (--miles=MI | --kilo=KM)

Available options:

--miles=MI

Distance in miles

--kilo=KM

Distance in kilometers

-h, --help

Prints help information

Users can pass value that satisfy either parser

$ app --miles 42
42.0

$ app --kilo 15
15.0

But not both at once or not at all:

$ app --miles 53 --kilo 10
Error: --kilo cannot be used at the same time as --miles

$ app
Error: expected --miles=MI or --kilo=KM, pass --help for usage information

If those cases are valid you can handle them with optional and many

If parsers inside parallel composition can parse the same object - the longest possible match should go first since bpaf picks an earlier parser if everything else is equal, otherwise it does not matter. In this example construct!([miles, km]) produces the same results as construct!([km, miles]) and only --help message is going to be different.

Parsers created with construct! still implement the Parser trait so you can apply more transformation on top. For example same as you can make a simple parser optional - you can make a composite parser optional. Parser transformed this way will succeed if both --alpha and --beta are present or neither of them:

struct Options {
    alpha: usize,
    beta: usize
}
 
fn parser() -> impl Parser<Option<Options>> {
    let alpha = long("alpha").argument("ALPHA");
    let beta = long("beta").argument("BETA");
    construct!(Options { alpha, beta }).optional()
}

← Transforming parsers

↑ Combinatoric API ↑

Subcommand parsers →