bpaf::_documentation::_1_tutorials::_1_combinatoric_api

Module _4_decorating

← Subcommand parsers

↑ Combinatoric API ↑

Improving the user experience

Once you have the final parser done there are still a few ways you can improve user experience. OptionParser comes equipped with a few methods that let you set version number, description, help message header and footer and so on.

#[derive(Debug, Clone)]
pub struct Options {
    argument: u32,
}

pub fn options() -> OptionParser<Options> {
    let argument = short('i').argument::<u32>("ARG");
    construct!(Options { argument })
        .to_options()
        .version("3.1415")
        .descr("This is a short description")
        .header("It can contain multiple blocks, this block goes before options")
        .footer("This one goes after")
}

fn main() {
    println!("{:?}", options().run())
}

Output

In addition to all the arguments specified by user bpaf adds a few more. One of them is --help:

$ app --help

This is a short description

Usage: app -i=ARG

It can contain multiple blocks, this block goes before options

Available options:

-i=ARG

-h, --help

Prints help information

-V, --version

Prints version information

This one goes after

The other one is --version - passing a string literal or something like env!("CARGO_PKG_VERSION") to get version from cargo directly usually works

$ app --version

Version: 3.1415

Other than that bpaf tries its best to provide a helpful error messages

$ app
Error: expected -i=ARG, pass --help for usage information

And if all parsers are satisfied run produces the result

$ app -i 10
Options { argument: 10 }

There are a few other things you can do:

group some of the primitive parsers into logical blocks for --help message with Parser::group_help

add tests to make sure important combinations are handled the way they are supposed to after any future refactors with OptionParser::run_inner

add a test to make sure that bpaf internal invariants are satisfied with OptionParser::check_invariants

generate user documentation in manpage and markdown formats with OptionParser::render_manpage and OptionParser::render_markdown

← Subcommand parsers

↑ Combinatoric API ↑