bpaf::_documentation::_1_tutorials::_2_derive_api

Module _0_intro

↑ Derive API tutorial ↑

Customizing flag and argument names →

Getting started with derive macro

Let’s take a look at a simple example

use bpaf::*;

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options)]
pub struct Options {
    /// A custom switch
    switch: bool,

    /// A custom argument
    argument: usize,
}

fn main() {
    println!("{:?}", options().run())
}

Output

bpaf generates a help message

$ app --help

Usage: app [--switch] --argument=ARG

Available options:

--switch

A custom switch

--argument=ARG

A custom argument

-h, --help

Prints help information

And two parsers. Numeric argument is required, boolean switch is optional and fall back value is false.

$ app --switch
Error: expected --argument=ARG, pass --help for usage information

$ app --switch --argument 42
Options { switch: true, argument: 42 }

$ app --argument 42
Options { switch: false, argument: 42 }

bpaf is trying hard to guess what you are trying to achieve just from the types so it will pick up types, doc comments, presence or absence of names, but it is possible to customize all of it, add custom transformations, validations and more.

↑ Derive API tutorial ↑

Customizing flag and argument names →