Function bpaf::batteries::verbose_by_slice

pub fn verbose_by_slice<T: Copy + 'static, const N: usize>(
    offset: usize,
    items: [T; N]
) -> impl Parser<T>

--verbose and --quiet flags with results choosen from a slice of values

Parameters specify an array of possible values and a default index

Usage

use bpaf::batteries::*;

#[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd)]
enum Level {
   Error,
   Warning,
   Info,
   Debug,
   Trace,
}

fn verbose() -> impl Parser<Level> {
    use Level::*;
    verbose_by_slice(2, [Error, Warning, Info, Debug, Trace])
}