Function bpaf::positional

pub fn positional(metavar: &'static str) -> Positional<String>

Positional argument in utf8 (String) encoding

For named flags and arguments ordering generally doesn’t matter: most programs would understand -O2 -v the same way as -v -O2, but for positional items order matters: in unix cat hello world and cat world hello would display contents of the same two files but in different order.

Important restriction

To parse positional arguments from a command line you should place parsers for all your named values before parsers for positional items and commands. In derive API fields parsed as positional items or commands should be at the end of your struct/enum. Same rule applies to parsers with positional fields or commands inside: such parsers should go to the end as well.

Use check_invariants in your test to ensure correctness.

For example for non positional non_pos and positional pos parsers

let valid = construct!(non_pos(), pos());
let invalid = construct!(pos(), non_pos());

bpaf panics during help generation unless if this restriction holds

Combinatoric usage

fn input() -> impl Parser<String> {
    positional("INPUT")
}

Derive usage

bpaf_derive converts fields in tuple-like structures into positional items

#[derive(Debug, Clone, Bpaf)]
struct Options(String);

positional and positional_os annotations also accept an optional metavar name

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(positional("INPUT"))]
    input: String,
}

See also positional_os - a simiar function