Trait bpaf::Parser

pub trait Parser<T> {
    fn many(self) -> ParseMany<Self>
   where
        Self: Sized,
    { ... }
    fn some(self, message: &'static str) -> ParseSome<Self>
   where
        Self: Sized + Parser<T>,
    { ... }
    fn optional(self) -> ParseOptional<Self>
   where
        Self: Sized + Parser<T>,
    { ... }
    fn parse<F, R, E>(self, f: F) -> ParseWith<T, Self, F, E, R>
   where
        Self: Sized + Parser<T>,
        F: Fn(T) -> Result<R, E>,
        E: ToString,
    { ... }
    fn map<F, R>(self, map: F) -> ParseMap<T, Self, F, R>
   where
        Self: Sized + Parser<T>,
        F: Fn(T) -> R + 'static,
    { ... }
    fn from_str<R>(self) -> ParseFromStr<Self, R>
   where
        Self: Sized + Parser<T>,
    { ... }
    fn guard<F>(self, check: F, message: &'static str) -> ParseGuard<Self, F>
   where
        Self: Sized + Parser<T>,
        F: Fn(&T) -> bool,
    { ... }
    fn fallback(self, value: T) -> ParseFallback<Self, T>
   where
        Self: Sized + Parser<T>,
    { ... }
    fn fallback_with<F, E>(self, fallback: F) -> ParseFallbackWith<T, Self, F, E>
   where
        Self: Sized + Parser<T>,
        F: Fn() -> Result<T, E>,
        E: ToString,
    { ... }
    fn hide(self) -> ParseHide<Self>
   where
        Self: Sized + Parser<T>,
    { ... }
    fn group_help(self, message: &'static str) -> ParseGroupHelp<Self>
   where
        Self: Sized + Parser<T>,
    { ... }
    fn to_options(self) -> OptionParser<T>
   where
        Self: Sized + Parser<T> + 'static,
    { ... }
}

Simple or composed argument parser

Overview

It’s best to think of an object implementing Parser trait as a container with a value inside that are composable with other Parser containers using construct! and the only way to extract this value is by transforming it to OptionParser with to_options and running it with run. At which point you either get your value out or bpaf would generate a message describing a problem (missing argument, validation failure, user requested help, etc) and the program would exit.

Values inside can be of any type for as long as they implement Debug, Clone and there’s no lifetimes other than static.

When consuming the values you usually start with Parser<String> or Parser<OsString> which you then transform into something that your program would actually use. it’s better to perform as much parsing and validation inside the Parser as possible so the program itself gets strictly typed and correct value while user gets immediate feedback on what’s wrong with the arguments they pass.

For example suppose your program needs user to specify a dimensions of a rectangle, with sides being 1..20 units long and the total area must not exceed 200 units square. A parser that consumes it might look like this:

#[derive(Debug, Copy, Clone)]
struct Rectangle {
    width: u32,
    height: u32,
}

fn rectangle() -> impl Parser<Rectangle> {
    let invalid_size = "Sides of a rectangle must be 1..20 units long";
    let invalid_area = "Area of a rectangle must not exceed 200 units square";
    let width = long("width")
        .help("Width of the rectangle")
        .argument("PX")
        .from_str::<u32>()
        .guard(|&x| 1 <= x && x <= 10, invalid_size);
    let height = long("height")
        .help("Height of the rectangle")
        .argument("PX")
        .from_str::<u32>()
        .guard(|&x| 1 <= x && x <= 10, invalid_size);
    construct!(Rectangle { width, height })
        .guard(|&r| r.width * r.height <= 400, invalid_area)
}

Derive specific considerations

Every method defined on this trait belongs to the postprocessing section of the field annotation. bpaf_derive would try to figure out what chain to use for as long as there’s no options changing the type: you can use fallback, fallback_with, guard, hide and group_help but not the rest of them.

#[derive(Debug, Clone, Bpaf)]
struct Options {
    // no annotation at all - `bpaf_derive` inserts implicit `argument` and `from_str`
    number_1: u32,

    // fallback isn't changing the type so `bpaf_derive` still handles it
    #[bpaf(fallback(42))]
    number_2: u32,

    // `bpaf_derive` inserts implicit `argument`, `optional` and `from_str`
    number_3: Option<u32>,

    // fails to compile: you need to specify a consumer, `argument` or `argument_os`
    // #[bpaf(optional)]
    // number_4: Option<u32>

    // fails to compile: you also need to specify how to go from String to u32
    // #[bpaf(argument("N"), optional)]
    // number_5: Option<u32>,

    // explicit consumer and a full postprocessing chain
    #[bpaf(argument("N"), from_str(u32), optional)]
    number_6: Option<u32>,
}

Provided Methods

fn many(self) -> ParseMany<Self>where
Self: Sized,

Consume zero or more items from a command line and collect them into Vec

many only collects elements that only consume something from the argument list.

Combinatoric usage:

fn numbers() -> impl Parser<Vec<u32>> {
    short('n')
        .argument("NUM")
        .from_str::<u32>()
        .many()
}

Derive usage:

bpaf would insert implicit many when resulting type is a vector

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short, argument("NUM"))]
    numbers: Vec<u32>
}

But it’s also possible to specify it explicitly, both cases renerate the same code. Note, since using many resets the postprocessing chain - you also need to specify from_str

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short, argument("NUM"), from_str(u32), many)]
    numbers: Vec<u32>
}

Example

$ app -n 1 -n 2 -n 3
// [1, 2, 3]

See also

some also collects results to a vector but requires at least one element to succeed

fn some(self, message: &'static str) -> ParseSome<Self>where
Self: Sized + Parser<T>,

Consume one or more items from a command line

Takes a string used as an error message if there’s no specified parameters

some only collects elements that only consume something from the argument list.

Combinatoric usage:

let numbers
    = short('n')
    .argument("NUM")
    .from_str::<u32>()
    .some("Need at least one number");

Derive usage

Since using some resets the postprocessing chain - you also need to specify from_str or similar, depending on your type

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short, argument("NUM"), from_str(u32), some("Need at least one number"))]
    numbers: Vec<u32>
}

Example

$ app
// fails with "Need at least one number"
$ app -n 1 -n 2 -n 3
// [1, 2, 3]

See also

many also collects results to a vector but succeeds with no matching values

fn optional(self) -> ParseOptional<Self>where
Self: Sized + Parser<T>,

Turn a required argument into optional one

optional converts any failure caused by missing items into is None and passes the remaining parsing failures untouched.

Combinatoric usage

fn number() -> impl Parser<Option<u32>> {
    short('n')
        .argument("NUM")
        .from_str::<u32>()
        .optional()
}

Derive usage

By default bpaf_derive would automatically use optional for fields of type Option<T>, for as long as it’s not prevented from doing so by present postprocessing options

#[derive(Debug, Clone, Bpaf)]
struct Options {
   #[bpaf(short, argument("NUM"))]
   number: Option<u32>
}

But it’s also possible to specify it explicitly, in which case you need to specify a full postprocessing chain which starts from from_str in this example.

#[derive(Debug, Clone, Bpaf)]
struct Options {
   #[bpaf(short, argument("NUM"), from_str(u32), optional)]
   number: Option<u32>
}

Example

$ app
// None
$ app -n 42
// Some(42)

fn parse<F, R, E>(self, f: F) -> ParseWith<T, Self, F, E, R>where
Self: Sized + Parser<T>,
F: Fn(T) -> Result<R, E>,
E: ToString,

Apply a failing transformation to a contained value

This is a most general of transforming parsers and you can express remaining ones terms of it: map, from_str and guard.

Examples given here are a bit artificail, to parse a value from string you should use from_str.

Combinatoric usage:

fn number() -> impl Parser<u32> {
    short('n')
        .argument("NUM")
        .parse(|s| u32::from_str(&s))
}

Derive usage:

parse takes a single parameter: function name to call. Function type should match parameter F used by parse in combinatoric API.

fn read_number(s: String) -> Result<u32, ParseIntError> {
    u32::from_str(&s)
}

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short, argument("NUM"), parse(read_number))]
    number: u32
}

Example

$ app -n 12
// 12
// fails with "Couldn't parse "pi": invalid numeric literal"

fn map<F, R>(self, map: F) -> ParseMap<T, Self, F, R>where
Self: Sized + Parser<T>,
F: Fn(T) -> R + 'static,

Apply a pure transformation to a contained value

A common case of parse method, exists mostly for convenience.

Combinatoric usage

fn number() -> impl Parser<u32> {
    short('n')
        .argument("NUM")
        .from_str::<u32>()
        .map(|v| v * 2)
}

Derive usage

fn double(num: u32) -> u32 {
    num * 2
}

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short, argument("NUM"), from_str(u32), map(double))]
    number: u32,
}

Example

$ app -n 21
// 42

fn from_str<R>(self) -> ParseFromStr<Self, R>where
Self: Sized + Parser<T>,

Parse stored String using FromStr instance

A common case of parse method, exists mostly for convenience.

Combinatoric usage

fn speed() -> impl Parser<f64> {
    short('s')
        .argument("SPEED")
        .from_str::<f64>()
}

Derive usage

By default bpaf_derive would use from_str for any time it’s not familiar with so you don’t need to specify anything

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short, argument("SPEED"))]
    speed: f64
}

But it’s also possible to specify it explicitly

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short, argument("SPEED"), from_str(f64))]
    speed: f64
}

Example

$ app -s pi
// fails with "Couldn't parse "pi": invalid float literal"
$ app -s 3.1415
// Version: 3.1415

See also

Other parsing and restricting methods include parse and guard. For transformations that can’t fail you can use map.

fn guard<F>(self, check: F, message: &'static str) -> ParseGuard<Self, F>where
Self: Sized + Parser<T>,
F: Fn(&T) -> bool,

Validate or fail with a message

If value doesn’t satisfy the constraint - parser fails with the specified error message.

Combinatoric usage

fn number() -> impl Parser<u32> {
    short('n')
        .argument("NUM")
        .from_str::<u32>()
        .guard(|n| *n <= 10, "Values greater than 10 are only available in the DLC pack!")
}

Derive usage

Unlike combinator counterpart, derive variant of guard takes a function name instead of a closure, mostly to keep thing clean. Second argument can be either a string literal or a constant name for a static str.

fn dlc_check(number: &u32) -> bool {
    *number <= 10
}

const DLC_NEEDED: &str = "Values greater than 10 are only available in the DLC pack!";

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(short, argument("NUM"), guard(dlc_check, DLC_NEEDED))]
    number: u32,
}

Example

$ app -n 100
// fails with "Values greater than 10 are only available in the DLC pack!"
$ app -n 5
// 5

fn fallback(self, value: T) -> ParseFallback<Self, T>where
Self: Sized + Parser<T>,

Use this value as default if value isn’t present on a command line

Parser would still fail if value is present but failure comes from some transformation

Combinatoric usage

fn number() -> impl Parser<u32> {
    short('n')
        .argument("NUM")
        .from_str::<u32>()
        .fallback(42)
}

Derive usage

Expression in parens should have the right type, this example uses u32 literal, but it can also be your own type if that’s what you are parsing, it can also be a function call.

#[derive(Debug, Clone, Bpaf)]
struct Options {
   #[bpaf(short, argument("NUM"), from_str(u32), fallback(42))]
   number: u32
}

Example

$ app -n 100
// 10
$ app
// 42
$ app -n pi
// fails with "Couldn't parse "pi": invalid numeric literal"

See also

fallback_with would allow to try to fallback to a value that comes from a failing computation such as reading a file.

fn fallback_with<F, E>(self, fallback: F) -> ParseFallbackWith<T, Self, F, E>where
Self: Sized + Parser<T>,
F: Fn() -> Result<T, E>,
E: ToString,

Use value produced by this function as default if value isn’t present

Would still fail if value is present but failure comes from some earlier transformation

Combinatoric usage

fn username() -> impl Parser<String> {
    long("user")
        .argument("USER")
        .fallback_with::<_, Box<dyn std::error::Error>>(||{
            let output = std::process::Command::new("whoami")
                .stdout(std::process::Stdio::piped())
                .spawn()?
                .wait_with_output()?
                .stdout;
            Ok(std::str::from_utf8(&output)?.to_owned())
        })
}

Derive usage

fn get_current_user() -> Result<String, Box<dyn std::error::Error>> {
    let output = std::process::Command::new("whoami")
        .stdout(std::process::Stdio::piped())
        .spawn()?
        .wait_with_output()?
        .stdout;
    Ok(std::str::from_utf8(&output)?.to_owned())
}

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(long, argument("USER"), fallback_with(get_current_user))]
    user: String,
}

Example

$ app --user bobert
// "bobert"
$ app
// "pacak"

See also

fallback implements similar logic expect that failures aren’t expected.

fn hide(self) -> ParseHide<Self>where
Self: Sized + Parser<T>,

Ignore this parser during any sort of help generation

Best used for optional parsers or parsers with a defined fallback, usually for implementing backward compatibility or hidden aliases

Combinatoric usage

/// bpaf would accept both `-W` and `-H` flags, but the help message
/// would contain only `-H`
fn rectangle() -> impl Parser<(u32, u32)> {
    let width = short('W')
        .argument("PX")
        .from_str::<u32>()
        .fallback(10)
        .hide();
    let height = short('H')
        .argument("PX")
        .from_str::<u32>()
        .fallback(10)
        .hide();
    construct!(width, height)
}

Example

$ app -W 12 -H 15
// (12, 15)
$ app -H 333
// (10, 333)
$ app --help
// contains -H but not -W

Derive usage

#[derive(Debug, Clone, Bpaf)]
struct Rectangle {
    #[bpaf(short('W'), argument("PX"), from_str(u32), fallback(10), hide)]
    width: u32,
    #[bpaf(short('H'), argument("PX"), from_str(u32))]
    height: u32,
}

Example

$ app -W 12 -H 15
// Rectangle { width: 12, height: 15 }
$ app -H 333
// Rectangle { width: 10, height: 333 }
$ app --help
// contains -H but not -W

fn group_help(self, message: &'static str) -> ParseGroupHelp<Self>where
Self: Sized + Parser<T>,

Attach help message to a complex parser

bpaf inserts the group help message before the block with all the fields from the inner parser and an empty line after the block.

Combinatoric usage

fn rectangle() -> impl Parser<(u32, u32)> {
    let width = short('w')
        .argument("PX")
        .from_str::<u32>();
    let height = short('h')
        .argument("PX")
        .from_str::<u32>();
    construct!(width, height)
        .group_help("Takes a rectangle")
}

Example

$ app --help
<skip>
            Takes a rectangle
   -w <PX>  Width of the rectangle
   -h <PX>  Height of the rectangle

<skip>

Derive usage

#[derive(Debug, Clone, Bpaf)]
struct Rectangle {
    width: u32,
    height: u32,
}

#[derive(Debug, Clone, Bpaf)]
struct Options {
    #[bpaf(external, group_help("Takes a rectangle"))]
    rectangle: Rectangle
}

fn to_options(self) -> OptionParser<T>where
Self: Sized + Parser<T> + 'static,

Transform Parser into OptionParser to attach metadata and run

Combinatoric usage

fn parser() -> impl Parser<u32> {
    short('i')
        .argument("ARG")
        .from_str::<u32>()
}

fn option_parser() -> OptionParser<u32> {
    parser()
        .to_options()
        .version("3.1415")
        .descr("This is a description")
}

See OptionParser for more methods available after conversion.

Derive usage

Add a top level options annotation to generate OptionParser instead of default Parser.

In addition to options annotation you can also specify either version or version(value) annotation. Former uses version from cargo, later uses the specified value which should be an expression of type &'static str, see version.

#[derive(Debug, Clone, Bpaf)]
#[bpaf(options, version("3.1415"))]
/// This is a description
struct Options {
   verbose: bool,
}

Example

$ app --version
// Version: 3.1415
$ app --help
<skip>
This is a description
<skip>